SEO: make /json-formatter the interactive formatter
Resolves the formatter intent collision: /json-formatter was a Chrome-
extension landing page while ranking ~pos 7 for "json formatter"
(4.78M impressions, ~0.7% CTR).

- /json-formatter now renders the interactive format+validate tool
  (reuses JsonValidatorWrapper), with an H1, formatter/beautifier-focused
  title + description, self-canonical, and descriptive cross-links to
  pretty-print, tree viewer, minify, and the extension.
- Extension landing moved to /json-formatter/chrome-extension (its own
  title, canonical, and a link back to the online tool).
